These are the changes needed for MSVC2017.

+++ include/mbgl/style/sources/geojson_source.hpp
@@ -15,7 +15,7 @@ struct GeoJSONOptions {
     // GeoJSON-VT options
     uint8_t minzoom = 0;
     uint8_t maxzoom = 18;
-    uint16_t tileSize = util::tileSize;
+    uint16_t tileSize = uint16_t(util::tileSize);
     uint16_t buffer = 128;
     double tolerance = 0.375;
